Modelo

+++ HashRotas.h
@@ -28,14 +28,14 @@ namespace HashRotas
     {
     public:
 
-        u_int32_t tamTabela;
+        u_int64_t tamTabela;
         std::vector<HashNo*> *tabelaHash;
         int numRotas;
 
         HashRotas(int numClientes);
         ~HashRotas();
-        u_int32_t getHash(Solucao::Veiculo *veiculo);
-        u_int32_t getHash(Solucao::ClienteRota *clienteRota, const int tam, const int tipo);
+        u_int64_t getHash(Solucao::Veiculo *veiculo);
+        u_int64_t getHash(Solucao::ClienteRota *clienteRota, const int tam, const int tipo);
 
         bool getVeiculo(Solucao::ClienteRota *clienteRota, const int tam, const int tipo, double *poluicao, double *combustivel);
         HashNo* getVeiculo(Solucao::Veiculo *veiculo);
